Refactor so that `git info` can use hosted repository.
